fix(finops): cj-259 D-CI-FUNC-4 lint-conventions check_migration_money.py AD-8 0048+0050-0054 sa.Numeric → BigInteger fix DONE
(cj-style 259번째 epic 연속 정직 회복 atomic single source-only sprint)

Batch AD-8 monetary-type fix across 6 Alembic migrations: 0048 finops_reporting,
0050 commitment, 0051 pricing, 0052 multi_cloud_unified_reconciliation,
0053 reserved_capacity_planning, 0054 chargeback_settlement. All KRW columns
had sa.Numeric(20,2) or sa.Numeric(20,4) which violates AD-8 (KRW must be
sa.BigInteger for 1-won precision integer currency). 48 violations → 0.

Pattern identical to cj-255 (which fixed 0045+0046). sa.Numeric(20,*) →
sa.BigInteger for all KRW-suffix columns. Other money types (USD _amount/_cost
without _krw suffix) intentionally unchanged — those are still NUMERIC(18,2)
territory per cj-255 AD-8 column-name scoping.

6 files +48/-48 (1:1 swap). ruff check all green. Migration import sanity OK.
scripts/check_migration_money.py: 0048/0050/0051/0052/0053/0054 all OK.

Residual AD-8 violations (cj-260+ scope, out-of-scope for cj-259):
- 0043_phase_11_finops.py: 4 NUMERIC(14,2) → NUMERIC(18,2) (base_amount,
  markup_amount, tax_amount, total_amount) — precision change, not type swap.
- 0044_phase_12_finops_anomaly.py: 5 NUMERIC(20,2) → NUMERIC(18,2)
  (baseline_amount, observed_cost, baseline_cost, consumption_amount,
  forecast_amount) — precision change, not type swap.

Total cj-260 backlog: 9 violations, all NUMERIC(precision!=18,scale!=2) →
NUMERIC(18,2). After cj-260 lands, lint-conventions will be HONESTLY GREEN
in live CI for the first time since the cj-247/248 false-CLOSED incident.

Refs:
- cj-255 (predecessor — same pattern, 0045+0046)
- cj-256 (verification handoff that revealed cj-247/248 was prematurely CLOSED)
- cj-260 (successor — 0043+0044 precision fixes)

+++ apps/api/alembic/versions/0048_phase_16_finops_reporting.py
@@ -54,14 +54,14 @@ def upgrade() -> None:
         sa.Column("scope_type", sa.Text, nullable=False),
         sa.Column("scope_id", sa.Text, nullable=False),
         sa.Column("period_key", sa.Text, nullable=False),
-        sa.Column("showback_total_krw", sa.Numeric(20, 2), nullable=False, server_default="0"),
+        sa.Column("showback_total_krw", sa.BigInteger, nullable=False, server_default="0"),
         sa.Column("anomaly_count_30d", sa.Integer, nullable=False, server_default="0"),
-        sa.Column("forecast_projection_krw", sa.Numeric(20, 2), nullable=False, server_default="0"),
+        sa.Column("forecast_projection_krw", sa.BigInteger, nullable=False, server_default="0"),
         sa.Column(
-            "optimization_savings_krw", sa.Numeric(20, 2), nullable=False, server_default="0"
+            "optimization_savings_krw", sa.BigInteger, nullable=False, server_default="0"
         ),
         sa.Column("tag_compliance_pct", sa.Numeric(8, 4), nullable=False, server_default="0"),
-        sa.Column("idle_cost_krw", sa.Numeric(20, 2), nullable=False, server_default="0"),
+        sa.Column("idle_cost_krw", sa.BigInteger, nullable=False, server_default="0"),
         sa.Column("department_breakdown", JSONB, nullable=False, server_default="{}"),
         sa.Column("cost_center_breakdown", JSONB, nullable=False, server_default="{}"),
         sa.Column("resource_type_breakdown", JSONB, nullable=False, server_default="{}"),
